LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/window-domain-failure.https.sub-expected.txt


FAIL SharedArrayBuffer and a same-origin-domain (but not same-origin) iframe assert_equals: expected false but got true
